Fix camera extra settings menu
If the list of settings entries is scrolled, the adapter returns a
reused View instead of creating a new one. This makes the last
entries unusable. This has been broken for ages but only shows up
if the camera has more than a few settings (which the stock camera
doesn't).

Ported from CM10.

Change-Id: I076202bd228b76c82af65a06f94134c46ef7ac59